pyproject.toml
[tool.poetry]
name = "smart-schedule-nsau"
version = "0.1.0"
description = "Chat bot with class schedule for Novosibirsk State Agrarian University (NSAU)"
authors = ["Алексей Лесков <alexleskov2000@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.0"
aiohttp = "^3.8.4"
classic-components = ">=0.0.1,<0.1.0"
pydantic = "^1.10.7"
python-json-logger = "^2.0.7"
attrs = ">=21.2.0,<21.3.0"
aiofiles = "^23.1.0"
xlrd = "^2.0.1"
sqlalchemy = ">=1.4.23,<1.5.0"
click = "^8.1.3"
alembic = "^1.10.2"
python-dateutil = "^2.8.2"
psycopg2 = "^2.9.5"
classic-sql-storage = "^0.0.1"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
yapf = "^0.32.0"
flake8 = "^6.0.0"
toml = "^0.10.2"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
known_first_party = "resty"
line_length = 80


[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 80
allow_split_before_dict_value = false
dedent_closing_brackets = true
split_complex_comprehension = true
split_arguments_when_comma_terminated = true
blank_line_before_nested_class_or_def = true
split_before_dict_set_generator = true
split_before_first_argument = true
split_before_arithmetic_operator = true
split_before_logical_operator = true
split_before_bitwise_operator = true


[tool.flake8]
max-line-length = 80
select = ['F', 'E', 'W', ]
ignore = ['W503', ]
per-file-ignores = [
    '*/adapters/*/repositories.py:E711,E712',
    '*/adapters/*/repositories/*:E711,E712',
    '*/migrations/*:E711,E712',
    '*/alembic/versions/*:E711,E712',
    '__init__.py:F401',
]
count = true